% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islandcodes-package.R
\docType{package}
\name{islandcodes-package}
\alias{islandcodes}
\alias{islandcodes-package}
\title{Reference data and helpers for small island states and territories}
\description{
Provides a curated reference list of countries and territories with
classifications for Small Island Developing States (SIDS), sub-national
island jurisdictions (SNIJ), World Bank region and income group, and
political association. Sub-sovereign cases such as Aruba, Curacao,
Bonaire, Sint Maarten, the French overseas territories, and Aaland
Islands are represented with disambiguating codes that standard
country-code packages often collapse or omit.
}
\details{
The package is deliberately small. It exposes:
\itemize{
  \item the bundled \code{\link{islands}} dataset,
  \item predicates \code{\link{is_sids}}, \code{\link{is_snij}},
        \code{\link{is_small_island}},
  \item the joiner \code{\link{add_island_cols}}, and
  \item the filter shorthand \code{\link{small_islands}}.
}

It imports \pkg{countrycode} so country names entered by the user are
coerced to ISO codes before lookup. The intent is to extend rather than
replace \pkg{countrycode}: pass country names in, get small-island
classifications out.
}
\section{Source data}{

The bundled dataset is mirrored from
\url{https://github.com/University-of-Aruba/island-research-reference-data}
under CC BY 4.0. Run \code{citation("islandcodes")} for the canonical
references.
}

\section{Acknowledgments}{

With thanks to Edward Cheung, the Aruban engineer at NASA Goddard
who designed A.R.U.B.A. (the ASCS/NCS Relay Unit Breaker Assembly)
installed on the Hubble Space Telescope during the 2002 servicing
mission. Cheung engineered the backronym so children in his home
country would have something concrete to point to on the telescope.
The \code{\link{aruba}} function in this package is a small homage
in the same style: A.R.U.B.A. as Annotated Reference for Under-coded
Border Areas.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/University-of-Aruba/islandcodes}
  \item \url{https://university-of-aruba.github.io/islandcodes/}
  \item Report bugs at \url{https://github.com/University-of-Aruba/islandcodes/issues}
}

}
\author{
\strong{Maintainer}: Rendell de Kort \email{rendell.dekort@ua.aw} (\href{https://orcid.org/0009-0006-5402-0500}{ORCID})

Other contributors:
\itemize{
  \item  University of Aruba (Digital Competence Dutch Caribbean (DCDC) Network) [copyright holder, funder]
}

}
\keyword{internal}
